requirements.txt
PyYAML>=6.0
requests>=2.25.0
urllib3>=1.26.0
